requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
openai==1.14.1
nba_api==1.4.1
pandas==2.2.1
requests==2.31.0